requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
python-dateutil>=2.8.0
pydantic>=2.0.0

# Trend data collection
pytrends>=4.9.0

# Data processing & storage
pandas>=2.0.0
scipy>=1.9.0
duckdb>=0.9.0
numpy>=1.24.0
statsmodels>=0.14.0
prophet>=1.1
mcp>=1.0.0

# Optional: For YouTube API
# google-api-python-client>=2.100.0
# google-auth>=2.23.0
# google-auth-oauthlib>=1.1.0
# google-auth-httplib2>=0.1.1

# Retry logic
tenacity>=8.2.0
pybreaker>=1.0.0
radar-core>=0.2.0
structlog>=23.2.0
playwright>=1.40.0
